refactor: tests package & factories module
